refactor: move queries parsing implementation to is own file
