+++ src/sql/optimiser.rs
@@ -6,23 +6,23 @@ use crate::db::SqlError;
 use crate::sql::statement::{BinaryOperator, Expression, Statement, UnaryOperator, Value};
 use crate::vm::expression::resolve_only_expression;
 
-use super::statement::Insert;
+use super::statement::{Insert, Select, Update};
 
 pub(crate) fn optimise(statement: &mut Statement) -> Result<(), SqlError> {
     match statement {
-        Statement::Select {
+        Statement::Select(Select {
             columns,
             r#where,
             order_by,
             ..
-        } => {
+        }) => {
             simplify_iter(columns.iter_mut())?;
             simplify_where(r#where)?;
             simplify_iter(order_by.iter_mut())?;
         }
-        Statement::Update {
+        Statement::Update(Update {
             columns, r#where, ..
-        } => {
+        }) => {
             simplify_where(r#where)?;
             simplify_iter(columns.iter_mut().map(|col| &mut col.value))?;
         }
