+++ src/sql/parser/mod.rs
@@ -8,6 +8,7 @@
 //! It's mostly inspired by the way postgres'
 //! [parser stage](https://www.postgresql.org/docs/17/parser-stage.html) works.
 
+mod queries;
 mod tokenizer;
 mod tokens;
 
@@ -22,7 +23,7 @@ use std::iter::Peekable;
 use tokenizer::{Location, TokenWithLocation, Tokenizer, TokenizerError};
 use tokens::{Keyword, Token};
 
-use super::statement::Insert;
+use super::statement::{Insert, Select, Update};
 
 pub(crate) struct Parser<'input> {
     input: &'input str,
@@ -79,38 +80,10 @@ impl<'input> Parser<'input> {
 
     pub fn parse_statement(&mut self) -> ParserResult<Statement> {
         let statement = match self.expect_one(&Self::supported_statements())? {
-            Keyword::Select => {
-                let cols = self.parse_separated_tokens(|parser| parser.parse_expr(None), false)?;
-                self.expect_keyword(Keyword::From)?;
-                let (from, r#where) = self.parse_from_and_where()?;
-
-                let order_by = self.parse_order_by()?;
-
-                Statement::Select {
-                    columns: cols,
-                    from,
-                    r#where,
-                    order_by,
-                }
-            }
+            Keyword::Select => Statement::Select(Select::parse(self)?),
             Keyword::Create => Statement::Create(Create::parse(self)?),
             Keyword::Insert => Statement::Insert(Insert::parse(self)?),
-            Keyword::Update => {
-                let table = self.parse_ident()?;
-                self.expect_keyword(Keyword::Set)?;
-
-                let columns = self.parse_separated_tokens(Self::parse_assign, false)?;
-                let r#where = match self.consume_optional(Token::Keyword(Keyword::Where)) {
-                    true => Some(self.parse_expr(None)?),
-                    false => None,
-                };
-
-                Statement::Update {
-                    columns,
-                    r#where,
-                    table,
-                }
-            }
+            Keyword::Update => Statement::Update(Update::parse(self)?),
             Keyword::Drop => {
                 let keyword = self.expect_one(&[Keyword::Database, Keyword::Table])?;
                 let identifier = self.parse_ident()?;
@@ -556,96 +529,6 @@ impl<'input> Parser<'input> {
     }
 }
 
-impl<'sql> Sql<'sql> for Create {
-    fn parse(parser: &mut Parser<'sql>) -> ParserResult<Self> {
-        let keyword = parser.expect_one(&[
-            Keyword::Database,
-            Keyword::Table,
-            Keyword::Unique,
-            Keyword::Index,
-        ])?;
-
-        Ok(match keyword {
-            Keyword::Database => Create::Database(parser.parse_ident()?),
-            Keyword::Table => Create::Table {
-                name: parser.parse_ident()?,
-                columns: parser.parse_separated_tokens(Parser::parse_col, true)?,
-            },
-            Keyword::Unique | Keyword::Index => {
-                let unique = keyword.eq(&Keyword::Unique);
-                if unique {
-                    parser.expect_keyword(Keyword::Index)?;
-                }
-
-                let name = parser.parse_ident()?;
-                parser.expect_keyword(Keyword::On)?;
-                let table = parser.parse_ident()?;
-                parser.expect_token(Token::LeftParen)?;
-                let column = parser.parse_ident()?;
-                parser.expect_token(Token::RightParen)?;
-
-                Create::Index {
-                    name,
-                    column,
-                    table,
-                    unique,
-                }
-            }
-            Keyword::Sequence => {
-                let name = parser.parse_ident()?;
-                parser.expect_keyword(Keyword::As)?;
-                // FIXME: hardcoded type
-                parser.expect_keyword(Keyword::BigInt)?;
-                parser.expect_keyword(Keyword::Unsigned)?;
-                parser.expect_keyword(Keyword::Owned)?;
-                parser.expect_keyword(Keyword::By)?;
-                let table = parser.parse_ident()?;
-
-                Create::Sequence {
-                    table,
-                    name,
-                    r#type: Type::UnsignedBigInteger,
-                }
-            }
-            _ => panic!("Unsupported keyword"),
-        })
-    }
-}
-
-impl<'sql> Sql<'sql> for Insert {
-    fn parse(parser: &mut Parser<'sql>) -> ParserResult<Self> {
-        parser.expect_keyword(Keyword::Into)?;
-        let into = parser.parse_ident()?;
-        let columns: Vec<String> = match parser.peek_token() {
-            Some(_) => parser.parse_separated_tokens(Parser::parse_ident, true),
-            None => Ok(vec![]),
-        }?;
-
-        parser.expect_keyword(Keyword::Values)?;
-
-        let mut rows = Vec::new();
-        loop {
-            // each row is ( expr, expr, ... )
-            parser.expect_token(Token::LeftParen)?;
-            let row = parser.parse_separated_tokens(|p| p.parse_expr(None), false)?;
-            parser.expect_token(Token::RightParen)?;
-            rows.push(row);
-
-            // comma? if so, more rows; otherwise break
-            match parser.consume_optional(Token::Comma) {
-                true => continue,
-                false => break,
-            }
-        }
-
-        Ok(Self {
-            into,
-            columns,
-            values: rows,
-        })
-    }
-}
-
 impl Display for ParserError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         writeln!(
@@ -743,7 +626,7 @@ mod tests {
 
         assert_eq!(
             statement,
-            Ok(Statement::Select {
+            Ok(Statement::Select(Select {
                 columns: vec![
                     Expression::Identifier("id".to_string()),
                     Expression::Identifier("price".to_string())
@@ -751,7 +634,7 @@ mod tests {
                 from: "bills".to_string(),
                 r#where: None,
                 order_by: vec![],
-            })
+            }))
         );
     }
 
@@ -762,7 +645,7 @@ mod tests {
 
         assert_eq!(
             statement,
-            Ok(Statement::Select {
+            Ok(Statement::Select(Select {
                 columns: vec![
                     Expression::Identifier("title".to_string()),
                     Expression::Identifier("author".to_string())
@@ -776,7 +659,7 @@ mod tests {
                     ))),
                 }),
                 order_by: vec![],
-            })
+            }))
         );
     }
 
@@ -787,12 +670,12 @@ mod tests {
 
         assert_eq!(
             statement,
-            Ok(Statement::Select {
+            Ok(Statement::Select(Select {
                 columns: vec![Expression::Wildcard],
                 from: "users".to_string(),
                 r#where: None,
                 order_by: vec![],
-            })
+            }))
         );
     }
 
@@ -881,7 +764,7 @@ mod tests {
 
         assert_eq!(
             statement,
-            Ok(Statement::Update {
+            Ok(Statement::Update(Update {
                 table: "employees".to_string(),
                 columns: vec![
                     Assignment {
@@ -928,7 +811,7 @@ mod tests {
                         }),
                     })))
                 }),
-            })
+            }))
         );
     }
 
@@ -999,12 +882,12 @@ mod tests {
                         },
                     ],
                 }),
-                Statement::Select {
+                Statement::Select(Select {
                     columns: vec![Expression::Wildcard],
                     from: "employees".to_string(),
                     r#where: None,
                     order_by: vec![],
-                }
+                })
             ]
         );
 
@@ -1020,7 +903,7 @@ mod tests {
 
         assert_eq!(
             statement,
-            Ok(Statement::Select {
+            Ok(Statement::Select(Select {
                 columns: vec![Expression::Wildcard],
                 from: "schedule".to_string(),
                 r#where: Some(Expression::BinaryOperation {
@@ -1029,7 +912,7 @@ mod tests {
                     right: Box::new(Expression::Value(Value::String("12:00:00".into()))),
                 }),
                 order_by: vec![],
-            })
+            }))
         );
     }
