+++ src/sql/query/planner.rs
@@ -10,7 +10,7 @@ use crate::{
     sql::{
         analyzer,
         query::optimiser,
-        statement::{Column, Expression, Type},
+        statement::{Column, Expression, Select, Type, Update},
         Statement,
     },
     vm::{
@@ -21,8 +21,8 @@ use crate::{
 use crate::{
     sql::statement::Insert,
     vm::planner::{
-        Collect, CollectBuilder, Delete, Project, SortBuilder, TupleComparator, Update,
-        DEFAULT_SORT_BUFFER_SIZE,
+        Collect, CollectBuilder, Delete, Project, SortBuilder, TupleComparator,
+        Update as UpdatePlan, DEFAULT_SORT_BUFFER_SIZE,
     },
 };
 
@@ -43,12 +43,12 @@ pub(crate) fn generate_plan<File: Seek + Read + Write + FileOperations>(
                 pager: Rc::clone(&db.pager),
             })
         }
-        Statement::Select {
+        Statement::Select(Select {
             columns,
             from,
             r#where,
             order_by,
-        } => {
+        }) => {
             let mut source = optimiser::generate_seq_plan(&from, r#where, db)?;
             let page_size = db.pager.borrow().page_size;
             let work_dir = db.work_dir.clone();
@@ -130,11 +130,11 @@ pub(crate) fn generate_plan<File: Seek + Read + Write + FileOperations>(
                 input: table.schema.clone(),
             })
         }
-        Statement::Update {
+        Statement::Update(Update {
             table,
             columns,
             r#where,
-        } => {
+        }) => {
             let mut source = optimiser::generate_seq_plan(&table, r#where, db)?;
             let work_dir = db.work_dir.clone();
             let page_size = db.pager.borrow().page_size;
@@ -149,7 +149,7 @@ pub(crate) fn generate_plan<File: Seek + Read + Write + FileOperations>(
                 }));
             }
 
-            Planner::Update(Update {
+            Planner::Update(UpdatePlan {
                 comparator: metadata.comp()?,
                 table: metadata.clone(),
                 assigments: columns,
